// src/app/core/enums/selector-item-type.ts
export enum SelectorItemType {
  BAFA = 'bafaLicense',
  CUSTOMER = 'customer',
  CUS_DELIVER_TO_ADDRESS = 'cusDeliverToAddress',
  CUS_SHIP_TO_ADDRESS = 'cusShipToAddress',
  CUS_ADDRESS_QUERY_ADDRESS ='customerAddressQueryForLicenseAddress',
  CUS_ADDRESS_QUERY_SHIP ='customerAddressQueryForLicenseShip',
  END_USER = 'endUser',
  HK_CONSIGNEE = 'hkConsignee',
  IMPORT_LICENSE_BY_FORMNO = 'importLicenseByFormNo',
  ITEM = 'item',
  NONE = 'none',
  OU = 'ou',
  OOU = 'oou',
  SHIP_FROM_ADDRESS = 'shipFromAddress',
  SPECIAL_IMPORT_LICENSE = 'specialImportLicesne',
  VENDOR = 'vendor',
  ALLCUSTOMER = 'allCustomer',
  ALLVENDOR = 'allVendor',
  ALLOU = 'allou',
  ALLOOU = 'alloou',
  CUSTOMER_INCLUDE_ALL = 'customer_all',
  VENDOR_INCLUDE_ALL = 'vendor_all',
  ALLOU_ALL = 'allou_all',
  ALLOOU_ALL = 'alloou_all',
  BRAND = 'brand',
  BRAND_WITHOUT_ALL = 'brand_without_all'
}